Update and rename testpython.py to getmetadatapython.py
Little python function to recover the metadata encoded in a .txt file. 

To use it, copy the html section of the article in a .txt file and write its name on the last line: searchmentions("myfile.txt")
On line 46, you can print any of the metadata you need, just be sure that it is the name contained in the class of your span right after mention. Ex: <span class="mention food" about="iceCream" id="ic1-22"> --> print(my_dict['food'])